requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
transformers==4.35.2
torch>=2.0.0
PyPDF2==3.0.1
python-docx==1.1.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
nltk>=3.8.0
spacy>=3.7.0
sentence-transformers>=2.2.0
accelerate>=0.24.0
